Move parse_skip_axis function to _utils module
